+++ tests/plugin-test/css-extract/normalizeUrl.test.js
@@ -1,4 +1,4 @@
-const normalizeUrl = require("../../../packages/rspack/dist/builtin-plugin/css-extract/hmr/normalize-url");
+const { normalizeUrl } = require("../../../packages/rspack/dist/builtin-plugin/css-extract/hmr/normalizeUrl");
 const dataUrls = require("./fixtures/json/data-urls.json");
 
 describe("normalize-url", () => {
